requirements.txt
fastapi==0.116.1
openai==1.98.0
python-dotenv==1.1.1
uvicorn==0.35.0
